moved clap.py to pam_utils folder
